python-api/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
supabase==2.8.1
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.41
